Ancillary arguments for controlling the outputs of coxph_mpl_dc

This is used to set various numeric parameters controlling a Cox model fit using coxph_mpl_dc. Typically it would only be used in a call to coxph_mpl_dc.

coxph_mpl_dc.control(ordSp,
                       binCount, tie,
                       tau, copula,
                       pent, smpart, penc, smparc,
                       maxit2, maxit,
                       mid, asy, ac, cv,
                       ac.theta, ac.gamma, ac.Utheta, ac.Ugamma,
                       min.theta, min.gamma,
                       min.ht, min.hc, min.St, min.Sc, min.C, min.dC,
                       eps, tol.thga, tol.bph, cat.smpar, tol.smpar
                       )

Arguments

ordSp
the order of spline for the basis function for baseline hazard for both T and C, can be 'piecewise constant' if ordSp=1, cubic 'm-spline' if ordSp=4, etc. Default is ordSp=1.

binCount
the number of subjects in each discretized bin, can be selected either by trial and error or AIC method Default is binCount=1.

tie
tie='No' if tied observations are not existed, otherwise tied observations existed. Default is tie='No'.

tau
the kendall’s correlation coefficient between T and C. Default is tau=0.

copula
Archimedean copula type, i.e. 'independent', 'clayton', 'gumbel' and 'frank'. Default is copula='independent'.

pent
penalty function type for T, i.e. mat1 (first order difference) or mat2 (second order difference) for piecewise constant basis, penalty_mspl for m-spline basis Default is pent='mat1'.

smpart
value of smoothing parameter for T, can be selected by either trial and error or cross validation method. Note that smpart can be also estimated by restricted maximum likelihood (i.e. smpart='REML'). Default is smpart=0.

penc
penalty function type for C, i.e. mat1 (first order difference) or mat2 (second order difference) for piecewise constant basis, penalty_mspl for m-spline basis Default is pent='mat1'.

smparc
value of smoothing parameter for C, can be selected by either trial and error or cross validation method. Note that smparc can be also estimated by restricted maximum likelihood (i.e. smparc='REML'). Default is smparc=0.

maxit2
maximum number of iterations for smpart and smparc. Defualt is maxit2=50.

maxit
maximum number of iteration for updating beta, phi, theta and gamma given fixed smpart and smparc. Default is maxit=5000.

mid
the middle matrix selection for the sandwich formula that used to computed the asymptotic covariance matrix, i.e. mid=1 (negative of the hessian matrix with zeros smoothing parameters, i.e. smpart=smparc=0, or negative of the matrix with second derivatives of the MPL estimates with respect to the log-likelihood), 2 (the matrix created by the vector of first derivative of the penalized log-likelihood with respect to the MPL estimates times its transpose) and otherwise (negative of the hessian matrix or negative of the matrix with second derivatives of the MPL estimates with respect to the penalized log-likelihood). Default is mid=1.

asy
asy=1 if asymptotic standard deviation of the MPL estimates are computed and 0 if not computed. Default is asy=1.

ac
ac=1 if aic value is calculated 0 if not. Default is ac=0.

cv
cv=0 if cv value is calculated 0 if not. Default is cv=0.

ac.theta
the minimum value of theta for active contraints. Default is ac.theta=1e-5.

ac.gamma
the minimum value of gamma for active contraints. Default is ac.gamma=1e-5.

ac.Utheta
the minimum value of Utheta (the first derivative of the penalized log-likelihood with respect to theta) for active contraints. Default is ac.Utheta=1e-2.

ac.Ugamma
the minimum value of Ugamma (the first derivative of the penalized log-likelihood with respect to gamma) for active contraints. Default is ac.Ugamma=1e-2.

min.theta
a value indicating the minimal baseline hazard parameter value theta updated at each iteration. Baseline hazard parameter theta estimates at each iteration lower than min.theta will be considered as min.theta. Default is min.theta=1e-7.

min.gamma
a value indicating the minimal baseline hazard parameter value gamma updated at each iteration. Baseline hazard parameter gamma estimates at each iteration lower than min.gamma will be considered as min.gamma. Default is min.gamma=1e-7.

min.ht
a value indicating the minimal baseline hazard of T updated at each iteration. Baseline hazard estimates of T at each iteration lower than min.ht will be considered as min.ht. Default is min.ht=1e-7.

min.hc
a value indicating the minimal baseline hazard of C updated at each iteration. Baseline hazard estimates of C at each iteration lower than min.hc will be considered as min.hc. Default is min.hc=1e-7.

min.St
a value indicating the minimal baseline survival of T updated at each iteration. Baseline survival estimates of T at each iteration lower than min.St will be considered as min.St. Default is min.St=1e-7.

min.Sc
a value indicating the minimal baseline survival of C updated at each iteration. Baseline survival estimates of C at each iteration lower than min.Sc will be considered as min.Sc. Default is min.Sc=1e-7.

min.C
a value indicating the minimal copula \(K(u,v)\) at each iteration, lower than min.C will be considered as min.C. Default is min.C=1e-7.

min.dC
a value indicating the minimal first i.e. \(dK(u,v)/du\) and \(dK(u,v)/dv\) and second i.e. \(d^2K(u,v)/dudv\) derivatives of copula \(K(u,v)\) at each iteration, lower than min.dC will be considered as min.dC. Default is min.dC=1e-7.

eps
a small positive value added to the diagonal of a square matrix. Default value is eps=1e-5.

tol.thga
the convergence tolerence value for both theta and gamma. Convergence is achieved when the maximum absolute difference between the parameter estimates at iteration k and iteration k-1 is smaller than tol.thga. Default is tol.thga=1e-5.

tol.bph
the convergence tolerence value for both beta and phi. Convergence is achieved when the maximum absolute difference between the parameter estimates at iteration k and iteration k-1 is smaller than tol.bph. Default is tol.bph=1e-5.

cat.smpar
cat.smpar='Yes' to display the smoothing parameters estimation process, otherwise not to display. Default is cat.smpar='Yes'.

tol.smpar
the convergence tolerence value for both smpart and smparc. Convergence is achieved when the maximum absolute difference between the parameter estimates at iteration k and iteration k-1 is smaller than tol.smpar. Default is tol.smpar=1e-2.

Value

A list containing the values of each of the above arguments for most of the inputs of Coxph_mpl_dc.

References

Ma, J. and Heritier, S. and Lo, S. (2014). "On the Maximum Penalised Likelihood Approach forProportional Hazard Models with Right Censored Survival Data". Computational Statistics and Data Analysis 74, 142-156.

Xu J, Ma J, Connors MH, Brodaty H. (2018). "Proportional hazard model estimation under dependent censoring using copulas and penalized likelihood". Statistics in Medicine 37, 2238–2251.

See also

Examples

control <- coxph_mpl_dc.control(ordSp=4,
                             binCount=40,
                             tau=0.8, copula='frank',
                             pent='penalty_mspl', smpart='REML', penc='penalty_mspl', smparc='REML',
                             cat.smpar='No'
                             )